+++ Nos/Controller/UNSWizardController.swift
@@ -37,7 +37,7 @@ class UNSWizardController: ObservableObject {
     }
     
     @Published var state: FlowState
-    @Published var authorKey: HexadecimalString?
+    @Published var authorKey: RawAuthorID?
     @Published var textField: String 
     @Published var phoneNumber: String?
     @Published var nameRecord: UNSNameRecord?
@@ -52,7 +52,7 @@ class UNSWizardController: ObservableObject {
     
     internal init(
         state: UNSWizardController.FlowState = .intro, 
-        authorKey: HexadecimalString? = nil, 
+        authorKey: RawAuthorID? = nil, 
         textField: String = "", 
         phoneNumber: String? = nil, 
         nameRecord: UNSNameRecord? = nil,