+++ NosTests/NoteParserTests.swift
@@ -11,17 +11,8 @@ import Dependencies
 
 // swiftlint:disable type_body_length
 
-final class NoteNoteParserTests: XCTestCase {
+final class NoteNoteParserTests: CoreDataTestCase {
 
-    @Dependency(\.persistenceController) private var persistenceController
-    var context: NSManagedObjectContext?
-    var continuation: DependencyValues.Continuation?
-
-    override func setUp() async throws {
-        persistenceController.resetForTesting()
-        context = persistenceController.viewContext
-    }
-    
     /// Example taken from [NIP-27](https://github.com/nostr-protocol/nips/blob/master/27.md)
     func testMentionWithNPub() throws {
         let mention = "@mattn"
@@ -43,7 +34,7 @@ final class NoteNoteParserTests: XCTestCase {
         let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
         let expectedContent = "hello @\(name)"
         let tags = [["p", hex]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let author = try Author.findOrCreate(by: hex, context: context)
         author.displayName = name
         try context.save()
@@ -78,7 +69,7 @@ final class NoteNoteParserTests: XCTestCase {
         let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
         let expectedContent = "@\(displayName)"
         let tags = [["p", hex]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
         let parsedContent = String(attributedContent.characters)
         XCTAssertEqual(parsedContent, expectedContent)
@@ -94,7 +85,7 @@ final class NoteNoteParserTests: XCTestCase {
         let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
         let expectedContent = "Hello\n@\(displayName)"
         let tags = [["p", hex]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
         let parsedContent = String(attributedContent.characters)
         XCTAssertEqual(parsedContent, expectedContent)
@@ -109,7 +100,7 @@ final class NoteNoteParserTests: XCTestCase {
         let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
         let expectedContent = "hello#[0]"
         let tags = [["p", hex]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
         let parsedContent = String(attributedContent.characters)
         XCTAssertEqual(parsedContent, expectedContent)
@@ -123,7 +114,7 @@ final class NoteNoteParserTests: XCTestCase {
         let content = "hello nostr:npub1937vv2nf06360qn9y8el6d8sevnndy7tuh5nzre4gj05xc32tnwqauhaj6"
         let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
         let tags = [["p", hex]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let author = try Author.findOrCreate(by: hex, context: context)
         author.displayName = name
         try context.save()
@@ -140,7 +131,7 @@ final class NoteNoteParserTests: XCTestCase {
         let content = "hello nostr:npub1937vv2nf06360qn9y8el6d8sevnndy7tuh5nzre4gj05xc32tnwqauhaj6"
         let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
         let tags = [["p", hex]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
         let links = attributedContent.links
         XCTAssertEqual(links.count, 1)
@@ -154,7 +145,7 @@ final class NoteNoteParserTests: XCTestCase {
         let content = "hello nostr:nprofile1qqszclxx9f5haga8sfjjrulaxncvkfekj097t6f3pu65f86rvg49ehqj6f9dh"
         let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
         let tags = [["p", hex]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let author = try Author.findOrCreate(by: hex, context: context)
         author.displayName = name
         try context.save()
@@ -170,7 +161,7 @@ final class NoteNoteParserTests: XCTestCase {
         let content = "Hello nostr:npub1pu3vqm4vzqpxsnhuc684dp2qaq6z69sf65yte4p39spcucv5lzmqswtfch.\n\nBye"
         let hex = "0f22c06eac1002684efcc68f568540e8342d1609d508bcd4312c038e6194f8b6"
         let tags = [["p", hex]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let author = try Author.findOrCreate(by: hex, context: context)
         author.displayName = name
         try context.save()
@@ -188,7 +179,7 @@ final class NoteNoteParserTests: XCTestCase {
         let displayName2 = "npub180cvv..."
         let hex2 = "3bf0c63fcb93463407af97a5e5ee64fa883d107ef9e558472c4eb9aaaefa459d"
         let tags = [["p", hex1], ["p", hex2]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
         let links = attributedContent.links
         XCTAssertEqual(links.count, 2)
@@ -203,7 +194,7 @@ final class NoteNoteParserTests: XCTestCase {
         let npub = "npub1937vv2nf06360qn9y8el6d8sevnndy7tuh5nzre4gj05xc32tnwqauhaj6"
         let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
         let tags: [[String]] = [[]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
         let links = attributedContent.links
         XCTAssertEqual(links.count, 1)
@@ -215,7 +206,7 @@ final class NoteNoteParserTests: XCTestCase {
         let content = "Check this note1h2mmqfjqle48j8ytmdar22v42g5y9n942aumyxatgtxpqj29pjjsjecraw"
         let hex = "bab7b02640fe6a791c8bdb7a352995522842ccb55779b21bab42cc1049450ca5"
         let tags: [[String]] = [[]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
         let links = attributedContent.links
         XCTAssertEqual(links.count, 1)
@@ -227,7 +218,7 @@ final class NoteNoteParserTests: XCTestCase {
         let content = "Check this nostr:note1h2mmqfjqle48j8ytmdar22v42g5y9n942aumyxatgtxpqj29pjjsjecraw"
         let hex = "bab7b02640fe6a791c8bdb7a352995522842ccb55779b21bab42cc1049450ca5"
         let tags: [[String]] = [[]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
         let links = attributedContent.links
         XCTAssertEqual(links.count, 1)
@@ -243,7 +234,7 @@ final class NoteNoteParserTests: XCTestCase {
         let profileHex = "0f22c06eac1002684efcc68f568540e8342d1609d508bcd4312c038e6194f8b6"
         let noteHex = "bab7b02640fe6a791c8bdb7a352995522842ccb55779b21bab42cc1049450ca5"
         let tags: [[String]] = [["p", profileHex]]
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
         let links = attributedContent.links
         XCTAssertEqual(links.count, 2)
@@ -261,7 +252,7 @@ final class NoteNoteParserTests: XCTestCase {
         let tags: [[String]] = [[]]
         
         let expectedContent = content
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
 
         let parsedContent = String(attributedContent.characters)
@@ -284,7 +275,7 @@ final class NoteNoteParserTests: XCTestCase {
         let tags: [[String]] = [[]]
 
         let expectedContent = "check this 🔗 Link to note"
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
 
         let parsedContent = String(attributedContent.characters)
@@ -307,7 +298,7 @@ final class NoteNoteParserTests: XCTestCase {
         let tags: [[String]] = [[]]
 
         let expectedContent = "check this 🔗 Link to note. Bye!"
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
 
         let parsedContent = String(attributedContent.characters)
@@ -328,7 +319,7 @@ final class NoteNoteParserTests: XCTestCase {
         let tags: [[String]] = [[]]
 
         let expectedContent = content
-        let context = try XCTUnwrap(context)
+        let context = try XCTUnwrap(testContext)
         let (attributedContent, _) = NoteParser.parse(content: content, tags: tags, context: context)
 
         let parsedContent = String(attributedContent.characters)