+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fixed an issue where bad data in the contact list could break the home feed.
 - Fixed a bug where the margins of root notes appeared incorrectly on Mac and iPad.
 - Fixed a date localization issue.
 - Optimized loading of the Notifications tab